+++ sidebar.php
@@ -1,9 +1,9 @@
-<div id="sidebar" class="col-lg-3 col-md-3 col-sm-12 col-xs-12">
+<div id="sidebar" class="col-lg-3 col-md-3 col-sm-12 col-xs-12 hidden-xs">
 	<?php if (is_active_sidebar('primary-sidebar')) {
         dynamic_sidebar('primary-sidebar');
     } else {
 	$args = array(
-		'before_widget' => '<div class="widget">',
+		'before_widget' => '<div id="%1$s" class="widget">',
         'after_widget' => '</div>',
         'before_title' => '<div class="title"><h2>',
         'after_title' => '</h2></div>'